requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.5.0
pydantic==1.10.2
PyYAML==6.0
